+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/guards/index.ts
@@ -1,1 +1,1 @@
-export * from './cancellation-return-request-input.guard';
+export * from './cancel-or-return-request-input.guard';
