+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/index.ts
@@ -1,3 +1,3 @@
-export * from './cancellation-return-items/cancellation-return-items.module';
+export * from './cancel-or-return-items/cancel-or-return-items.module';
 export * from './return-order/return-order.module';
 export * from './return-order-confirmation/return-order-confirmation.module';
