+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/return-order-confirmation/return-order-confirmation.module.ts
@@ -9,9 +9,9 @@ import {
 } from '@spartacus/core';
 import { PageLayoutComponent } from '../../../../../cms-structure/page/page-layout/page-layout.component';
 import { CmsPageGuard } from '../../../../../cms-structure/guards/cms-page.guard';
-import { CancelOrReturnItemsModule } from '../cancellation-return-items/cancellation-return-items.module';
+import { CancelOrReturnItemsModule } from '../cancel-or-return-items/cancel-or-return-items.module';
 import { ReturnOrderConfirmationComponent } from './return-order-confirmation.component';
-import { CancelOrReturnRequestInputGuard } from '../guards/cancellation-return-request-input.guard';
+import { CancelOrReturnRequestInputGuard } from '../guards/cancel-or-return-request-input.guard';
 
 @NgModule({
   imports: [
