+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/return-order/return-order.component.spec.ts
@@ -13,7 +13,7 @@ import { ReturnOrderComponent } from './return-order.component';
 
 @Component({
   template: '',
-  selector: 'cx-cancellation-return-items',
+  selector: 'cx-cancel-or-return-items',
 })
 class MockCancelOrReturnItemsComponent {
   @Input() entries: OrderEntry[];
